+++ lib/modules/settings/modules/configuration_radarr_headers.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'package:fluro/fluro.dart';
 import 'package:flutter/material.dart';
 import 'package:lunasea/core.dart';
+import 'package:lunasea/modules/radarr.dart';
 import 'package:lunasea/modules/settings.dart';
 
 class SettingsConfigurationRadarrHeadersRouter extends LunaPageRouter {
@@ -90,6 +91,7 @@ class _State extends State<_SettingsConfigurationRadarrHeadersRoute> {
                         _headers.addAll({'Authorization': 'Basic $_auth'});
                         Database.currentProfileObject.radarrHeaders = _headers;
                         Database.currentProfileObject.save();
+                        Provider.of<RadarrState>(context, listen: false).reset();
                     }
                 break;
                 case 100:
@@ -99,6 +101,7 @@ class _State extends State<_SettingsConfigurationRadarrHeadersRoute> {
                         _headers.addAll({results[1]: results[2]});
                         Database.currentProfileObject.radarrHeaders = _headers;
                         Database.currentProfileObject.save();
+                        Provider.of<RadarrState>(context, listen: false).reset();
                     }
                 break;
                 default: